+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+* Feat: Support for platform stacktraces on Android (#788)
+
 ## 6.5.0-alpha.2
 
 * Bump: Sentry-Android to 5.7.0 and Sentry-Cocoa to 7.11.0 (#796)